fix: keep docs generator as moved file
